# [flow] Generalize conv1x1 to matmul for NCHW