Most models from torch uses conv2d.nchw by default. To accelerate the 1x1 conv workload on nchw cases, we generalize the conv1x1 to matmul pass to be able to handle NCHW as well.

Perf boost on ResNet50 (torch):
CPU(Threadripper PRO 3995WX): 100ms -> 80ms
Vulkan( Threadripper PRO 3995WX): 48ms -> 31.7ms